only create zram if not exists

+++ scripts/sbin/setup_zram.sh
@@ -11,7 +11,9 @@ totalmem=$(free -m | awk '/Mem:/{print $2}')
 mem=$(( ($totalmem / 4 / 2)* 1024 * 1024))
 
 for i in `seq 0 3`; do
-  echo $mem > /sys/block/zram${i}/disksize
-  mkswap /dev/zram${i}
-  swapon -p 5 /dev/zram${i}
+  if [[ ! -e /dev/zram${i} ]]; then
+    echo $mem > /sys/block/zram${i}/disksize
+    mkswap /dev/zram${i}
+    swapon -p 5 /dev/zram${i}
+  fi
 done
